Add MetaMask Snap-based Encryption
This PR adds sample UI to use the built-in MetaMask "signing utils"
snap to encrypt and descrypt data.

1. Users can connect their MetaMask wallet and request their public
   encryption key
2. You can for any given public encryption key, generate an ecrypted
   blob that only that MetaMask wallet can decrypt

The UI is kind of janky ATM, but that is OK, since this is just to try
out the tech, and not meant for actual use. (As a note to self - I
should change the UI to allow encryption without connecting a wallet).