Add Passkey Encryption Scheme (#5)
This PR adds a new encruption scheme that uses the Passkey PRF extension
to generate some deterministic entropy that we use as a seed for an
encryption key pair.

This is very similar to the Ceramic DID scheme, but a slight improvement
as it:
- Uses a method that is **dedicated** for generating strong
cryptographic entropy, and is even designed to be used to generate
encryption keys (albeit symetric keys)
- Is not susceptible to phishing, as WebAuthn does not allow credentials
to be shared across relaying parties (i.e. Dapps in our case).

It still suffers from the same major downside though, that we are
materializing an encryption key in memory, so leaking it is really bad
(compared to MetaMask snap scheme, where you don't materialize the
decryption key stays in the wallet, meaning that you can at most
compromise the decrypted messages, and you maintain forwards and
backwards secrecy).

---------

Co-authored-by: Shebin John <admin@remedcu.com>